signoff: cell-pin SAIF expansion (99.5% of pins annotated), RTL module stripping, plans
OpenSTA's read_saif annotates pins, not nets: a net record is looked up as a
pin of its INSTANCE, so a flat netlist's nets must be written as the pins of
every leaf cell.  The simulator now traces nets only (cell internals untraced,
underscore nets included) and expand_saif_to_pins() writes them through the
routed netlist -- on the routed argmax collective 45,754 of 45,988 pins
annotate (the rest are synthesis temporaries whose names OpenSTA's SAIF lexer
rejects).  Instance names keep '.' and '$' raw, the only form it matches.

Replaced modules are stripped from their RTL file (helpers kept); plans can
take a campaign's own source list; Verilator 5 is preferred.

+++ tools/signoff/activity_harness.cpp
@@ -13,6 +13,7 @@
 #include "verilated.h"
 #include "verilated_vcd_c.h"
 #include <cstdio>
+#include <cstdint>
 #include <cstdlib>
 #include <cstring>
 #include <string>
@@ -50,7 +51,7 @@ int main(int argc, char** argv) {
         top->clk = !top->clk;
         top->eval();
         half++;
-        if (tfp) tfp->dump(half * hp);
+        if (tfp) tfp->dump(static_cast<uint64_t>(half * hp));
     }
     if (tfp) {
         tfp->close();
